Create ScaniaAnalysis
This was my first submission to Kaggle on the data from Scania trucks. I used a random forest aggregator and predicted that a checkup is needed if the predict value is strictly greater than zero. This seems to catch all the "pos" cases in the test set but does have around 926/16e3 false positives ~8%. Two heuristics were used to reduce the data set. 
1. A column of data with more than 10% 'na' values was elimated 2. Any column with more than 90% zeros, I checked if any of the non zero values corresponded with the class being 'pos' if any such column had zeros for all cases with class being pos, the column was eliminated

Comments & suggestions welcome. Thanks for reading.